requirements.txt
datasets==2.18.0
flash_attn==2.5.3
huggingface_hub==0.20.2
llama.egg==info
matplotlib==3.7.2
networkx==3.2.1
networkx_metis==1.0
numpy==1.24.3
openai==0.28.0
pandas==2.2.1
peft==0.9.0
pytorch_lightning==2.2.0.post0
pytorch_partial_crf==0.2.1
scikit_learn==1.3.1
scikit_learn==1.3.0
scipy==1.12.0
sentence_transformers==2.2.2
sqlitedict==2.1.0
torch==1.13.1
tqdm==4.65.0
transformers==4.38.2
